Rename API controllers/routes
Include the purpose of the file in the name, to help with finding and
loading a specific file.
